ftp/diskinfo/icon/library: fix AROS x86_64 format-string warnings
Users compiling dopus5 natively on AROS (reported on EAB) hit a build
failure at source/Modules/ftp/ftp_main.c:3322, where AROS' stricter
gcc rejects:

    D(bug("** ABORT %ld listers...\n", og->og_listercount));

On AROS x86_64 (ABIv11) int is 32-bit but long is 64-bit, and ULONG
expands to unsigned int (not unsigned long), so every legacy "%ld for
int" or "%lx for ULONG/pointer" turns into a -Wformat= error. Classic
m68k Amiga compiles these fine because LONG/ULONG/long are all 32-bit
there.

Audit all format strings in the main AROS build. In 12 files, 53 call
sites were affected; the fix matches the existing pattern in
ftp_module.c: keep the %ld/%lx/%lu specifiers and cast the argument
to (long) / (unsigned long). That keeps the output identical on
m68k/PPC/MorphOS while matching gcc's strict format checks on AROS
x86_64.

While auditing, three latent bugs surfaced (all in D(bug(...)) paths,
i.e. DEBUG-only but would crash/garble output when DEBUG is enabled):

  * ftp_main.c:800 had no argument at all for its "0x%lx" specifier;
    now prints the failed handle.
  * ftp.c:2213 passed a full struct in_addr to a %lx, instead of its
    s_addr field.
  * ftp_recursive.c:3180/3193 used %s for an unsigned int offset;
    switched to %lu, matching the neighbouring "REST %lu" FTP command.

Files touched:
  - Library/wb.c                         (1 site)
  - Modules/diskinfo/diskinfo.c          (1 site, 5 args)
  - Modules/icon/icon.c                  (2 sites)
  - Modules/ftp/ftp.c                    (5 sites)
  - Modules/ftp/ftp_addressbook.c        (3 sites)
  - Modules/ftp/ftp_addrsupp.c           (1 site)
  - Modules/ftp/ftp_lister.c             (4 sites)
  - Modules/ftp/ftp_lister_connect.c     (6 sites)
  - Modules/ftp/ftp_lister_xfer.c        (3 sites)
  - Modules/ftp/ftp_main.c               (9 sites)
  - Modules/ftp/ftp_recursive.c          (17 sites)
  - Modules/ftp/ftp_util.c               (1 site)

Verified clean builds (no format warnings, no errors) with:
  sacredbanana/amiga-compiler:m68k-amigaos   (os3)          OK
  midwan/aros-compiler:i386-aros             (i386-aros)    OK
  midwan/aros-compiler:x86_64-aros           (x86_64-aros)  OK

+++ source/Library/wb.c
@@ -2609,10 +2609,10 @@ BOOL LIBFUNC L_WB_Remove_Patch(REG(a6, struct MyLibrary *libbase))
 				// Get library
 				if ((libptr = wb_get_patchbase(wb_patches[patch].type, (struct LibData *)libbase->ml_UserData)))
 				{
-					D(bug("waiting for the usecount (%d) to be zero (patch type %d at LVO %d)\n",
+					D(bug("waiting for the usecount (%d) to be zero (patch type %d at LVO %ld)\n",
 						  usecount[patch],
 						  wb_patches[patch].type,
-						  wb_patches[patch].offset));
+						  (long)wb_patches[patch].offset));
 					while (usecount[patch] != 0)
 						Delay(1);
 					Forbid();
